\name{drawlogaxis} 
\Rdversion{1.1} 
\alias{drawlogaxi} 
\title{Draw a 10-log scale axis}
\description{}
\usage{ 
drawlogaxis(side,range)
}

\arguments{ 
\item{side}{The side of the plot (same as \kbd{axis})} 
\item{range}{The axis range, in power of ten}
}

\note{This function is used internally by the package}