\name{logaxis} 
\Rdversion{1.1} 
\alias{logaxis} 
\title{Draw a 10-log scale axis}
\description{Add a 10-log scale axis on the specified side of the plot}
\usage{ 
logaxis(side,range)
}

\arguments{ 
\item{side}{The side of the plot (same as \kbd{axis})} 
\item{range}{The axis range, in power of ten}
}

\note{This function is used internally by the package}